+++ app/src/main/java/com/codekotliners/memify/ui/viewmodels/ImageViewerViewModel.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.codekotliners.memify.core.ImageRepository
+import com.codekotliners.memify.core.model.ImageRepository
 import com.codekotliners.memify.core.model.ImageItem
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
